requirements.txt
matplotlib
pandas
langchain
langchain-community
chromadb
sentence-transformers
anthropic
openai
fastapi
symspellpy
pyyaml